+++ NvimView/Sources/NvimView/NvimView+Debug.swift
@@ -5,16 +5,16 @@
 
 import Cocoa
 
-extension NvimView {
-  @IBAction public func debug1(_: Any?) {
+public extension NvimView {
+  @IBAction func debug1(_: Any?) {
     do { try self.ugrid.dump() } catch { self.log.error("Could not dump UGrid: \(error)") }
   }
-  
-  @IBAction public func debug2(_: Any?) {
+
+  @IBAction func debug2(_: Any?) {
     self.log.debug("Nothing yet")
   }
 
-  func draw(cellGridIn context: CGContext) {
+  internal func draw(cellGridIn context: CGContext) {
     context.saveGState()
     defer { context.restoreGState() }