+++ RxPack/Sources/RxNeovim/RxNeovimApi.swift
@@ -40,25 +40,12 @@ public final class RxNeovimApi {
 
   public func stop() -> Completable { self.msgpackRpc.stop() }
 
-  public func checkBlocked<T>(_ single: Single<T>) -> Single<T> {
-    self
-      .getMode()
-      .flatMap { dict -> Single<T> in
-        guard (dict["blocking"]?.boolValue ?? false) == false else {
-          throw RxNeovimApi.Error.blocked
-        }
-
-        return single
-      }
-  }
-
-  public func rpc(
+  public func sendRequest(
     method: String,
-    params: [RxNeovimApi.Value],
-    expectsReturnValue: Bool = true
+    params: [RxNeovimApi.Value]
   ) -> Single<RxNeovimApi.Value> {
     self.msgpackRpc
-      .request(method: method, params: params, expectsReturnValue: expectsReturnValue)
+      .request(method: method, params: params, expectsReturnValue: true)
       .map { response -> RxMsgpackRpc.Value in
         guard response.error.isNil else { throw RxNeovimApi.Error(response.error) }