+++ VimR/VimR/AppDelegate.swift
@@ -20,9 +20,9 @@ final class UpdaterDelegate: NSObject, SPUUpdaterDelegate {
 
   func feedURLString(for _: SPUUpdater) -> String? {
     if self.useSnapshotChannel {
-      return "https://raw.githubusercontent.com/qvacua/vimr/master/appcast_snapshot.xml"
+      "https://raw.githubusercontent.com/qvacua/vimr/master/appcast_snapshot.xml"
     } else {
-      return "https://raw.githubusercontent.com/qvacua/vimr/master/appcast.xml"
+      "https://raw.githubusercontent.com/qvacua/vimr/master/appcast.xml"
     }
   }
 }